requirements.txt
streamlit
matplotlib
pillow
requests